feat(Probes): Adds the basic preset scan answers to the probe query response core. Closes #2361
